関数(getとrun）の共通化
